fix: libla.h prototype and conflicting types libla can compile now

+++ src/libla/matrix/cofactor/mat3_cofactor.c
@@ -12,9 +12,9 @@
 
 #include "libla.h"
 
-float	mat3_cofactor(const t_mat3 *m, int row, int col)
+float	mat3_cofactor(const t_mat3 *m, short row, short col)
 {
-	int	sign;
+	short	sign;
 
 	sign = 1;
 	if ((row + col) % 2 == 1)
