+++ src/libla/matrix/cofactor/mat4_cofactor.c
@@ -12,9 +12,9 @@
 
 #include "libla.h"
 
-float	mat4_cofactor(const t_mat4 *m, int row, int col)
+float	mat4_cofactor(const t_mat4 *m, short row, short col)
 {
-	int	sign;
+	short	sign;
 
 	sign = 1;
 	if ((row + col) % 2 == 1)
